requirements.txt
dask_geopandas==0.3.1
gensim==4.2.0
geopandas==0.13.0
matplotlib==3.6.2
mwparserfromhell==0.6.4
nltk==3.7
numpy==1.24.3
pandas==1.4.4
pickle5==0.0.12
Pillow==9.4.0
Pillow==10.0.0
pyproj==3.4.1
scikit_learn==1.2.1
seaborn==0.12.2
Shapely==2.0.1
smart_open==5.2.1
torch==2.0.1
torchmetrics==0.11.2
tqdm==4.65.0
transformers==4.24.0
wandb==0.15.2
